move module-specific types to module's types.ts file
