+++ src/app/modules/topas_app/topas_app_module.ts
@@ -1,6 +1,6 @@
 import { AfterGenesisBlockApplyContext, BaseModule, codec } from 'lisk-sdk';
 
-import { ModuleId, ModuleName, TopasApp, TopasAppModuleChainData } from '../../types';
+import { ModuleId, ModuleName } from '../../types';
 import { serializeData } from '../../utils/formats';
 import { getDataAccessData } from '../../utils/store';
 import { CreateAppAsset } from './assets/create_app_asset';
@@ -10,6 +10,7 @@ import { TipCreatorAsset } from './assets/tip_creator_asset';
 import { UpdateAppAsset } from './assets/update_app_asset';
 import { TOPAS_APP_MODULE_INIT, TOPAS_APP_MODULE_KEY } from './constants';
 import { topasAppAccountSchema, topasAppModuleSchema } from './schemas';
+import { TopasApp, TopasAppModuleChainData } from './types';
 
 export class TopasAppModule extends BaseModule {
 	public name = ModuleName.TopasApp;
